+++ api/src/main/java/org/allaymc/api/entity/interfaces/EntityProjectile.java
@@ -1,0 +1,10 @@
+package org.allaymc.api.entity.interfaces;
+
+import org.allaymc.api.entity.Entity;
+import org.allaymc.api.entity.component.EntityProjectileBaseComponent;
+
+/**
+ * @author daoge_cmd
+ */
+public interface EntityProjectile extends Entity, EntityProjectileBaseComponent {
+}