+++ api/src/main/java/org/allaymc/api/entity/interfaces/EntitySnowball.java
@@ -1,8 +1,5 @@
 package org.allaymc.api.entity.interfaces;
 
-import org.allaymc.api.entity.Entity;
-import org.allaymc.api.entity.component.EntityProjectileBaseComponent;
-
-public interface EntitySnowball extends Entity, EntityProjectileBaseComponent {
+public interface EntitySnowball extends EntityProjectile {
 
 }