+++ api/src/main/java/org/allaymc/api/eventbus/event/entity/ProjectileHitEvent.java
@@ -1,0 +1,38 @@
+package org.allaymc.api.eventbus.event.entity;
+
+import lombok.Getter;
+import org.allaymc.api.block.dto.BlockStateWithPos;
+import org.allaymc.api.entity.Entity;
+import org.allaymc.api.entity.interfaces.EntityProjectile;
+import org.allaymc.api.eventbus.event.CancellableEvent;
+
+/**
+ * @author daoge_cmd
+ */
+@Getter
+public class ProjectileHitEvent extends EntityEvent implements CancellableEvent {
+
+    /**
+     * The block being hit, can be {@code null}.
+     */
+    protected BlockStateWithPos block;
+    /**
+     * The entity being hit, can be {@code null}.
+     */
+    protected Entity victim;
+
+    public ProjectileHitEvent(EntityProjectile entity, BlockStateWithPos block) {
+        super(entity);
+        this.block = block;
+    }
+
+    public ProjectileHitEvent(EntityProjectile entity, Entity victim) {
+        super(entity);
+        this.victim = victim;
+    }
+
+    @Override
+    public EntityProjectile getEntity() {
+        return (EntityProjectile) super.getEntity();
+    }
+}